+++ R/toPresentation.R
@@ -12,13 +12,56 @@
 #'
 #' @seealso \code{\link{render.sigr_significance}}, \code{\link{render.sigr_ftest}}
 #' @export
-render <- function(statistic,...,
+render <- function(statistic,
+                   ...,
                    format,
                    statDigits=2,
                    sigDigits=2,
                    pLargeCutoff=0.05,
                    pSmallCutoff=1.0e-5) {
   UseMethod('render')
+}
+
+# get render args out of ...
+get_render_args <- function(args, strict, stat, fnname) {
+  res <- list(
+    # format = "ascii", # if commented out trigger format special case
+    statDigits=2,
+    sigDigits=2,
+    pLargeCutoff=0.05,
+    pSmallCutoff=1.0e-5
+  )
+  for(ni in names(res)) {
+    # first from stat, then override from args
+    if(ni %in% names(stat)) {
+      res[[ni]] <- stat[[ni]]
+    }
+    if(ni %in% names(args)) {
+      res[[ni]] <- args[[ni]]
+      args[[ni]] <- NULL
+    }
+  }
+  # if format not set, look for in stat and then environment
+  if(!("format" %in% names(res))) {
+    # format is a special case if not set in initial res
+    if("format" %in% names(args)) {
+      res$format <- args$format
+      args$format <- NULL
+    } else if("format" %in% names(stat)) {
+      res$format <- stat$format
+    } else {
+      res$format <- getRenderingFormat()
+    }
+  }
+  if(!isTRUE(res$format %in% formats)) {
+    res$format <- "ascii"
+  }
+  if(strict) {
+    if(length(args)>0) {
+      stop(paste("unexpected arguments", fnname))
+    }
+  }
+  res
 }
 
 
@@ -45,15 +88,13 @@ render.sigr_significance <- function(statistic,
                           sigDigits=2,
                           pLargeCutoff=0.05,
                           pSmallCutoff=1.0e-5) {
-  if(length(list(...))>0) {
-    stop("render.sigr_significance unexpected arguments")
-  }
+  wrapr::stop_if_dot_args(substitute(list(...)), "sigr::render.sigr_significance")
   sig <- statistic$significance
   symbol <- statistic$symbol
   if (missing(format) || is.null(format)) {
     format <- getRenderingFormat()
   }
-  if(!(format %in% formats)) {
+  if(!isTRUE(format %in% formats)) {
     format <- "ascii"
   }
   fsyms <- syms[format,]
@@ -74,38 +115,53 @@ render.sigr_significance <- function(statistic,
   pString
 }
 
+
 #' as.character
 #'
 #' @param x sigr wrapper to print
-#' @param ... extra arguments (not used)
+#' @param ... extra arguments for sigr::render
 #' @return formatted string
+#'
 #' @examples
 #'
 #' as.character(wrapSignificance(1/300))
 #'
 #' @export
 as.character.sigr_statistic <- function(x, ...) {
-  render(x,format='ascii')
+  render_args <- get_render_args(list(...), TRUE, x, "sigr::as.character.sigr_statistic")
+  render(x,
+         format = render_args$format,
+         statDigits = render_args$statDigits,
+         sigDigits = render_args$sigDigits,
+         pLargeCutoff = render_args$pLargeCutoff,
+         pSmallCutoff = render_args$pSmallCutoff)
 }
 
 #' Format
 #'
 #' @param x sigr wrapper to print
-#' @param ... extra arguments (not used)
+#' @param ... extra arguments for sigr::render
 #' @return formatted string
+#'
 #' @examples
 #'
 #' format(wrapSignificance(1/300))
 #'
 #' @export
 format.sigr_statistic <- function(x, ...) {
-  render(x,format='ascii')
+  render_args <- get_render_args(list(...), TRUE, x, "sigr::format.sigr_statistic")
+  render(x,
+         format = render_args$format,
+         statDigits = render_args$statDigits,
+         sigDigits = render_args$sigDigits,
+         pLargeCutoff = render_args$pLargeCutoff,
+         pSmallCutoff = render_args$pSmallCutoff)
 }
 
 #' Print
 #'
 #' @param x sigr wrapper to print
-#' @param ... extra arguments
+#' @param ... extra arguments for sigr::render and print
 #' @return formatted string
 #' @examples
 #'
@@ -113,7 +169,14 @@ format.sigr_statistic <- function(x, ...) {
 #'
 #' @export
 print.sigr_statistic <- function(x, ...) {
-  print(render(x,format='ascii'), ...)
+  render_args <- get_render_args(list(...), FALSE, x, "sigr::print.sigr_statistic")
+  print(render(x,
+               format = render_args$format,
+               statDigits = render_args$statDigits,
+               sigDigits = render_args$sigDigits,
+               pLargeCutoff = render_args$pLargeCutoff,
+               pSmallCutoff = render_args$pSmallCutoff),
+        ...)
 }
 
 #' Wrap a significance